Move tournament match functions
