+++ backend/scandamus/game/tournament_match.py
@@ -4,12 +4,23 @@ import logging
 
 from django.contrib.auth.models import User
 from .models import Player, Entry
-from .models import Match, Tournament
+from .models import Match, Tournament, Entry
 from django.conf import settings
 from django.db import transaction
 from .match_utils import authenticate_token, get_player_by_user
 from asgiref.sync import async_to_sync
 from channels.db import database_sync_to_async
+
+
+#import random
+
+#from celery import shared_task
+#from django.utils import timezone
+#from datetime import timedelta
+
+from channels.layers import get_channel_layer
+
+from .match_utils import send_tournament_match_jwt, notify_bye_player
 
 
 logger = logging.getLogger(__name__)
@@ -89,3 +100,150 @@ async def send_to_client(websocket, message):
     except Exception as e:
         logger.error(f'Failed to send to consumer: {e}')
         return
+    
+def report_match_result(match_id):
+    match = Match.objects.get(id=match_id)
+    tournament = match.tournament
+    current_round = match.round
+    logger.info(f'report_match_result in round:{current_round}')
+
+    matched_in_round = tournament.matches.filter(round=current_round)
+    if current_round in [-1, -3]: # round -1:決勝戦, -3:3位決定戦
+        final_match = tournament.matches.get(round=-1)
+        third_place_match = tournament.matches.get(round=-3)
+        if final_match.status == 'after' and third_place_match.status == 'after':
+            finalize_tournament(tournament)
+    elif current_round in [-4, -5, -6]: # 3人決戦
+        handle_three_players_round(tournament, current_round)
+    elif all(m.status == 'after' for m in matched_in_round):
+        create_next_round(tournament, current_round)
+
+# 最後にABCの3人が残った場合
+# 1戦目（round==-4）: A vs B
+# 2戦目（round==-5）: 1戦目の敗者 vs C　（敗者は3位）
+# 3戦目（round==-6）: 1戦目の勝者 vs 2戦目の勝者　（勝者は1位、敗者は2位）
+def handle_three_players_round(tournament, current_round):
+    logger.info('handle_three_players_round')
+    previous_round_match = tournament.matches.filter(round=current_round).order_by('-id').first()
+    if current_round == -4:
+        first_loser = previous_round_match.player1 if previous_round_match.winner == previous_round_match.player2 else previous_round_match.player2
+        create_match(tournament, first_loser, tournament.bye_player, -5)
+        tournament.bye_player = None
+        tournament.save()
+    elif current_round == -5:
+        first_match = tournament.matches.filter(round=-4).order_by('-id').first()
+        second_match = tournament.matches.filter(round=-5).order_by('-id').first()
+        tournament.third_place = second_match.player1 if second_match.winner == second_match.player2 else second_match.player2
+        tournament.save()
+        create_match(tournament, first_match.winner, second_match.winner, -6)
+    elif current_round == -6:
+        finalize_tounrnament_by_three_players(tournament)
+
+def finalize_tounrnament_by_three_players(tournament):
+    logger.info('finalize_tournament_by_three_players in')
+    final_match = tournament.matches.filter(round=-6).order_by('-id').first()
+
+    tournament.winner = final_match.winner
+    tournament.second_place = final_match.player1 if final_match.winner == final_match.player2 else final_match.player2
+    tournament.status = 'finished'
+    tournament.save()
+    tournament.finalize_result_json(True)
+
+def finalize_tournament(tournament):
+    logger.info('finalize_tournament in')
+    final_match = tournament.matches.filter(round=-1).order_by('-id').first()
+    third_place_match = tournament.matches.filter(round=-3).order_by('-id').first()
+
+    tournament.winner = final_match.winner
+    tournament.second_place = final_match.player1 if final_match.winner == final_match.player2 else final_match.player2
+    tournament.third_place = third_place_match.winner
+    tournament.status = 'finished'
+    tournament.save()
+    tournament.finalize_result_json()
+
+def create_next_round(tournament, current_round):
+    logger.info('create_next_round in')
+    previous_round_matches = tournament.matches.filter(round=current_round)
+    winners = [match.winner for match in previous_round_matches if match.winner]
+    tournament.update_result_json(current_round)
+
+    if tournament.bye_player:
+        winners.insert(0, tournament.bye_player)
+        tournament.bye_player == None
+        tournament.save()
+    
+    number_of_winners = len(winners)
+    if number_of_winners == 2:
+        create_final_round(tournament, winners, previous_round_matches)
+        return
+    elif number_of_winners == 3:
+        create_match(tournament, winners[0], winners[1], -4) # -4:3人決戦の1戦目
+        tournament.bye_player = winners[-1]
+        tournament.save()
+        notify_bye_player(tournament)#3人準決勝待ち TODO: 特別なメッセージ？
+        return
+    
+    current_round += 1
+
+    create_matches(tournament, winners, current_round)
+
+    tournament.current_round = current_round
+    tournament.save()
+
+def create_final_round(tournament, winners, previous_round_matches):
+    logger.info('create_final_round in')
+    # 決勝戦
+    create_match(tournament, winners[0], winners[1], round=-1)
+
+    semifinal_losers = [
+        match.player1 if match.winner == match.player2 else match.player2 for match in previous_round_matches
+    ]
+    # 3位決定戦
+    create_match(tournament, semifinal_losers[0], semifinal_losers[1], round=-3)
+
+def create_match(tournament, player1, player2, round, game_name='pong'):
+    match = Match.objects.create(
+        tournament=tournament,
+        player1=player1,
+        player2=player2,
+        round=round,
+        game_name=game_name,
+        status='before'
+    )
+    match.save()
+    tournament.matches.add(match)
+    tournament.save()
+    async_to_sync(send_tournament_match_jwt)(match)
+
+def create_matches(tournament, players, round_number):
+    number_of_players = len(players)
+    logger.info(f'create_matches in : number_of_players:{number_of_players}')
+    matches = []
+
+    for i in range(0, len(players) - 1, 2):
+        player1 = players[i]
+        player2 = players[i + 1]
+        match = Match(
+            tournament=tournament,
+            round=round_number,
+            player1=player1,
+            player2=player2,
+            status='before'
+        )
+        match.save()
+        matches.append(match)
+        async_to_sync(send_tournament_match_jwt)(match)
+
+    if len(players) % 2 == 1:
+        tournament.bye_player = players[-1] # 最後の要素のPlayer
+    else:
+        tournament.bye_player = None
+
+    logger.info(f'Type of matches: {type(matches)}')  # 型を確認
+    number_of_matches = len(matches)
+    logger.info(f'size of macthes: {number_of_matches}')
+   #Match.objects.bulk_create(matches)
+    tournament.matches.add(*matches)
+    tournament.save()
+    if tournament.bye_player:
+        notify_bye_player(tournament)
